pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-features"
version = "0.1.0"
description = "Audio-focused advertisement performance prediction system"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AudioAd Team", email = "team@audioad.ai"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio",
]
requires-python = ">=3.9,<3.13"
dependencies = [
    # Core ML (minimal set for content_based_predictor.py)
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    
    # Audio processing (minimal set)
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
    "ffmpeg-python>=0.2.0",
    
    # Speech recognition & NLP (minimal set)
    "speechrecognition>=3.10.0",
    "textblob>=0.17.0",
    "nltk>=3.8.0",
    
    # Utilities
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

cloud = [
    "google-cloud-speech>=2.20.0",
    "azure-cognitiveservices-speech>=1.32.0",
    "boto3>=1.28.0",  # AWS Transcribe
]

viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "jupyter>=1.0.0",
]

api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=audio_features",
    "--cov-report=term-missing",
] 

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
